[Update] 인공지능 서버와 연결

+++ src/main/java/com/example/bami/short_travel/dto/ShortTravelDTO.java
@@ -29,7 +29,7 @@ public class ShortTravelDTO {
 
     @Data
     public static class Location {
-        private double latitude;
-        private double longitude;
+        private float latitude;
+        private float longitude;
     }
 }
