+++ src/main/java/com/example/bami/short_travel/entity/TravelPlanEntity.java
@@ -28,10 +28,10 @@ public class TravelPlanEntity {
     private String endDate;  // 여행 종료 날짜
 
     @Column(nullable = false)
-    private Double latitude;
+    private float latitude;
 
     @Column(nullable = false)
-    private Double longitude;
+    private float longitude;
 
     private String address;
 
@@ -50,7 +50,7 @@ public class TravelPlanEntity {
         this.endDate = endDate;
     }
 
-    public void setLocation(Double latitude, Double longitude, String address) {
+    public void setLocation(float latitude, float longitude, String address) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.address = address;
